% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(df)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
remove unwanted features from the dataframe
mostly used to femove basic features and fun_evals and costs_runtime features,
but also features that produce missing or invalid values
}
