% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{clean_data_full}
\alias{clean_data_full}
\title{clean_data_full
remove unwanted features, and also remove all constant features}
\usage{
clean_data_full(features_cec_orig)
}
\arguments{
\item{features_cec_orig}{the dataframe to clean}
}
\value{
a dataframe with no constant and unwanted features
}
\description{
clean_data_full
remove unwanted features, and also remove all constant features
}
