% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{do_visualization}
\alias{do_visualization}
\title{do_visualization}
\usage{
do_visualization(
  name1,
  name2,
  save_prefix,
  step,
  nfun,
  dim,
  scale_x_factor = 10000,
  scale_y_factor = 10000,
  shift_x_factor = 10000,
  shift_y_factor = 10000,
  perplexity = 5,
  normalize_Y = FALSE,
  exclude = c(),
  error_on_fnf = FALSE,
  gecco = FALSE,
  gecco_cec = FALSE,
  saved_wilcoxon = FALSE,
  save_prefix_gecco = "",
  rotate = FALSE,
  nfun1 = 0,
  nfun2 = 0,
  cec_all = FALSE
)
}
\arguments{
\item{name1}{the filename containing the features from the first set of problems}

\item{name2}{the filename containing the features from the second set of problems}

\item{save_prefix}{}

\item{step}{}

\item{nfun}{}

\item{dim}{}

\item{scale_x_factor}{}

\item{scale_y_factor}{}

\item{shift_x_factor}{}

\item{shift_y_factor}{}

\item{perplexity}{}

\item{normalize_Y}{}

\item{exclude}{}

\item{error_on_fnf}{}

\item{gecco}{}

\item{gecco_cec}{}

\item{saved_wilcoxon}{}

\item{save_prefix_gecco}{}

\item{rotate}{}

\item{nfun1}{}

\item{nfun2}{}

\item{cec_all}{}
}
\value{

}
\description{
preforms all steps needed to visualize the provided data
}
