% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{filter_features}
\alias{filter_features}
\title{filter_features
remove coulmns containing the specified features from df}
\usage{
filter_features(df, features)
}
\arguments{
\item{df}{the dataframe to remove columns from}

\item{features}{the vector of column names to remove}
}
\value{

}
\description{
filter_features
remove coulmns containing the specified features from df
}
