% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{generate_data_2}
\alias{generate_data_2}
\title{generate_data_2
take in two sets of problem samples from files name1 and name2, and return a combined dataframe
containing combined samples, with samples in name2 shifted and scaled}
\usage{
generate_data_2(
  root_folder = "C:\\\\R_Code\\\\Landscape Analysis\\\\rezultati\\\\samples",
  name1,
  name2,
  dim,
  step,
  nfun,
  scale_x_factor = 10000,
  scale_y_factor = 10000,
  shift_x_factor = 10000,
  shift_y_factor = 10000,
  normalize_Y = FALSE,
  gecco = FALSE,
  gecco_cec = FALSE,
  rotate = FALSE
)
}
\arguments{
\item{root_folder}{the path of the root folder containing all of the sample files}

\item{name1}{filename of the first problem sample file, without the .mat extension}

\item{name2}{filename of the second problem sample file, without the .mat extension}

\item{dim}{the dimension of the data}

\item{step}{number of samples for each individual problem}

\item{nfun}{total number of problems}

\item{scale_x_factor}{how much to scale the x1,...,xn part of the problem}

\item{scale_y_factor}{how much to scale the y part of the problem}

\item{shift_x_factor}{how much to shift the x1,...,xn part of the problem}

\item{shift_y_factor}{how much to shift the y part of the problem}

\item{normalize_Y}{should y valus be normalized}

\item{gecco}{do the files contain gecco problems (FALSE for cec problems)}

\item{gecco_cec}{does one of the files contain gecco problems, and the other CEC problems (FALSE if both files contain the same benchmark)}

\item{rotate}{should the problems also be rotated (not used in the paper, as rotation wasn't considered)}
}
\value{
a combined dataframe containing samples and results from name1 and name2, with the samples and results in name2 shifted and scaled
}
\description{
generate_data_2
take in two sets of problem samples from files name1 and name2, and return a combined dataframe
containing combined samples, with samples in name2 shifted and scaled
}
