% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize the features in a dataframe}
\usage{
normalize_data(df)
}
\arguments{
\item{df}{the dataframe to normalize}
}
\value{
a normalized dataframe
}
\description{
Not used in the paper, but provided as an option
}
