% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{process_features}
\alias{process_features}
\title{process_features}
\usage{
process_features(
  data,
  save_prefix,
  nfun,
  errors,
  saved_wilcoxon = FALSE,
  wilcoxon_filename = "C:\\\\R_Code\\\\saved_wilcoxon.rds"
)
}
\arguments{
\item{data}{The dataframe of landscape features to preprocess}

\item{save_prefix}{prefix of the filename that will contained the saved data about which features are affected by shifting and scaling}

\item{nfun}{The total number of problems in the data}

\item{errors}{Indices of rows which contained errors during feature calculation}

\item{saved_wilcoxon}{Should we use existing knowledge of which features are affected by shifting and scaling, or calculate it from the provided data. If FALSE, the calculated features will be saved to a file.}

\item{wilcoxon_filename}{The filename containing the names of the features affected by shifting and scaling. Used when saved_wilcoxon = TRUE}
}
\value{
A processed dataframe
}
\description{
do the final feature processing. This includes
1. Removing all problems where the landscape feature calculation encountered an error
2. Removing the features affected by shifting and scaling
3. Removing correlated features
4. Normalizing the data
5. Performing PCA
}
\details{
The function allows two modes of operation. In the first, we calculate the features affected by shifting and scaling from the provided data.
In the second, we use an existing file which lists which features are affected.

In order to not overfit the data, we first calculated the affected features using only the data from the 2014 CEC problems, saved this data,
and then used these affected features when plotting comparisons between CEC and GECCO problems
}
